chess_moves_classification_model/requirements.txt
numpy==1.24.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
Pillow==9.4.0
